add label directly without computedLabel, add "required" style class, move tests
