+++ packages/vue-vanilla/tests/unit/controls/StringControlRenderer.spec.ts
@@ -13,6 +13,23 @@ const uischema = {
   },
 };
 
+const schemaRequired = {
+  type: 'object',
+  properties: {
+    a: {
+      type: 'string',
+    },
+  },
+  required: ['a'],
+};
+const uischemaRequired = {
+  type: 'Control',
+  scope: '#/properties/a',
+  options: {
+    hideRequiredAsterisk: true,
+  },
+};
+
 describe('StringControlRenderer.vue', () => {
   it('renders a string input', () => {
     const wrapper = mountJsonForms('a', schema, uischema);
@@ -36,5 +53,15 @@ describe('StringControlRenderer.vue', () => {
     const input = wrapper.find('input');
     const placeholder = input.attributes('placeholder');
     expect(placeholder).to.equal('string placeholder');
+  });
+
+  it('should have a lable with required class and asterisk symbol', () => {
+    const wrapper = mountJsonForms('a', schemaRequired, uischema);
+    expect(wrapper.find('label.required span.asterisk').exists()).to.be.true;
+  });
+
+  it('should have a lable with required class but asterisk symbol is hidden', () => {
+    const wrapper = mountJsonForms('a', schemaRequired, uischemaRequired);
+    expect(wrapper.find('label.required span.asterisk').exists()).to.be.false;
   });
 });
